Função InputBox adicionada ao AppUtils
